Move spatial reference and filters to sub directories.
