+++ include/sempr/entity/spatial/filter/LocalTransformationFilter.hpp
