+++ include/sempr/entity/spatial/reference/GeodeticCS.hpp
@@ -1,7 +1,7 @@
 #ifndef SEMPR_ENTITY_SPATIAL_GEODETICCS_HPP_
 #define SEMPR_ENTITY_SPATIAL_GEODETICCS_HPP_
 
-#include <sempr/entity/spatial/GlobalCS.hpp>
+#include <sempr/entity/spatial/reference/GlobalCS.hpp>
 //#include <ogr_spatialref.h>
 
 namespace sempr { namespace entity {
