+++ include/sempr/entity/spatial/reference/SpatialReference.hpp
