+++ src/entity/spatial/reference/GlobalCS.cpp
@@ -1,5 +1,5 @@
+#include <sempr/entity/spatial/reference/GlobalCS.hpp>
 #include <GlobalCS_odb.h>
-#include <sempr/entity/spatial/GlobalCS.hpp>
 
 #include <string>
 #include <typeinfo>
@@ -48,41 +48,24 @@ Eigen::Affine3d GlobalCS::transformationFromRoot() const
     return Eigen::Affine3d::Identity();
 }
 
-std::shared_ptr<geos::geom::CoordinateFilter> GlobalCS::to(const GlobalCS::Ptr other)
+FilterList GlobalCS::to(const GlobalCS::Ptr other)
 {
-    //no single transformatin of the reference is not the same and representation is different!
-
-/*
-    auto thisType = typeid(*this);
-    auto otherType = typeid(*other);
-    bool sameReference = this->sameReference(other);
-
-    if (thisType != otherType)
-    {
-        // tranformation between different types
-        if(sameReference)
-        {
-
-        }
-        else
-        {
-            // could not change the reference and transform it in one step!
-        }
-    }
-    else
-    {
-        // same type
-        if (!sameReference)
-        {
-            // reference transformatin
-        }
-        else
-        {
-            // nothing to do!
-        }
-    }*/
-
+    //only for ODB support
+    return FilterList();
 }
+
+std::shared_ptr<geos::geom::CoordinateFilter> GlobalCS::froward()
+{
+    //only for ODB support
+    return std::shared_ptr<geos::geom::CoordinateFilter>(nullptr);
+}
+
+std::shared_ptr<geos::geom::CoordinateFilter> GlobalCS::reverse()
+{
+    //only for ODB support
+    return std::shared_ptr<geos::geom::CoordinateFilter>(nullptr);
+}
+
 
 
 }}
